Use sparse representation of flags
